+++ src/Symfony/Component/Translation/Formatter/IntlMessageFormatter.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Translation\Formatter;
 
+use Symfony\Component\Translation\Exception\InvalidArgumentException;
+
 /**
  * @author Guilherme Blanco <guilhermeblanco@hotmail.com>
  * @author Abdellatif Ait boudad <a.aitboudad@gmail.com>
@@ -25,15 +27,12 @@ class IntlMessageFormatter implements MessageFormatterInterface
         try {
             $formatter = new \MessageFormatter($locale, $message);
         } catch (\Throwable $e) {
-            throw new \InvalidArgumentException('Invalid message format.', $e);
-        }
-        if (null === $formatter) {
-            throw new \InvalidArgumentException(sprintf('Invalid message format. Reason: %s (error #%d)', intl_get_error_message(), intl_get_error_code()));
+            throw new InvalidArgumentException(sprintf('Invalid message format. Reason: %s (error #%d)', intl_get_error_message(), intl_get_error_code()), 0, $e);
         }
 
         $message = $formatter->format($parameters);
         if (U_ZERO_ERROR !== $formatter->getErrorCode()) {
-            throw new \InvalidArgumentException(sprintf('Unable to format message. Reason: %s (error #%s)', $formatter->getErrorMessage(), $formatter->getErrorCode()));
+            throw new InvalidArgumentException(sprintf('Unable to format message. Reason: %s (error #%s)', $formatter->getErrorMessage(), $formatter->getErrorCode()));
         }
 
         return $message;